board/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
python-dotenv==1.0.0

